requirements.txt
Flask
flask-cors
firebase-admin
requests
Pillow
openai
python-dotenv
